Extract Pulse widget
